refactor: 채팅방 관련 클래스 디렉토리 이동
