+++ src/main/java/com/gobongbob/festamate/domain/chat/application/ChatService.java
@@ -1,11 +1,11 @@
 package com.gobongbob.festamate.domain.chat.application;
 
 import com.gobongbob.festamate.domain.chat.domain.Chat;
+import com.gobongbob.festamate.domain.chat.domain.ChatRoom;
 import com.gobongbob.festamate.domain.chat.dto.request.ChatRequest;
 import com.gobongbob.festamate.domain.chat.dto.response.ChatResponse;
 import com.gobongbob.festamate.domain.chat.persistence.ChatRepository;
-import com.gobongbob.festamate.domain.chatRoom.domain.ChatRoom;
-import com.gobongbob.festamate.domain.chatRoom.repository.ChatRoomRepository;
+import com.gobongbob.festamate.domain.chat.persistence.ChatRoomRepository;
 import com.gobongbob.festamate.domain.member.domain.Member;
 import com.gobongbob.festamate.domain.member.persistence.MemberRepository;
 import java.util.List;
